+++ packages/dsw-document-worker/CHANGELOG.md
@@ -8,6 +8,11 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ## [Unreleased]
 
 
+## [3.28.0]
+
+### Added
+
+- Send charset as part of S3 object content-type
 ## [3.27.1]
 
 ### Fixed
@@ -182,3 +187,4 @@ Released for version consistency with other DSW tools.
 [3.26.1]: /../../tree/v3.26.1
 [3.27.0]: /../../tree/v3.27.0
 [3.27.1]: /../../tree/v3.27.1
+[3.28.0]: /../../tree/v3.28.0